refactor: SetEventNamePage, SetEventPasswordPage 비즈니스 로직 hook으로 분리
