+++ client/src/pages/CreateEventPage/SetEventNamePage.tsx
@@ -1,40 +1,10 @@
-import {useState} from 'react';
-import {useNavigate} from 'react-router-dom';
 import {FixedButton, MainLayout, LabelInput, Title, TopNav, Back} from 'haengdong-design';
+import {css} from '@emotion/react';
 
-import validateEventName from '@utils/validate/validateEventName';
-
-import useEvent from '@hooks/useEvent';
-
-import {ROUTER_URLS} from '@constants/routerUrls';
+import useSetEventNamePage from '@hooks/useSetEventNamePage';
 
 const SetEventNamePage = () => {
-  const [eventName, setEventName] = useState('');
-  const [errorMessage, setErrorMessage] = useState('');
-  const [canSubmit, setCanSubmit] = useState(false);
-  const navigate = useNavigate();
-  const {createNewEvent} = useEvent();
-
-  const submitEventName = async (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-
-    navigate(ROUTER_URLS.eventCreatePassword, {state: {eventName}});
-  };
-
-  const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const newValue = event.target.value;
-    const validation = validateEventName(newValue);
-
-    setCanSubmit(newValue.length !== 0);
-
-    if (validation.isValid) {
-      setEventName(newValue);
-      setErrorMessage('');
-    } else {
-      event.target.value = eventName;
-      setErrorMessage(validation.errorMessage ?? '');
-    }
-  };
+  const {submitEventName, errorMessage, eventName, canSubmit, handleChange} = useSetEventNamePage();
 
   return (
     <MainLayout>
@@ -42,7 +12,7 @@ const SetEventNamePage = () => {
         <Back />
       </TopNav>
       <Title title="행사 이름 입력" description="시작할 행사 이름을 입력해 주세요." />
-      <form onSubmit={submitEventName} style={{padding: '0 1rem'}}>
+      <form onSubmit={submitEventName} css={css({padding: '0 1rem'})}>
         <LabelInput
           labelText="행사 이름"
           errorText={errorMessage}
